requirements.txt
absl-py==2.0.0
accelerate==0.16.0
aiohttp==3.8.5
aiohttp-cors==0.7.0
aiosignal==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
astor==0.8.1
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
autogluon==0.8.2
autogluon.common==0.8.2
autogluon.core==0.8.2
autogluon.features==0.8.2
autogluon.multimodal==0.8.2
autogluon.tabular==0.8.2
autogluon.timeseries==0.8.2
autograd==1.6.2
autograd-gamma==0.5.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
blessed==1.20.0
blinker==1.6.2
blis==0.7.11
boto3==1.28.73
botocore==1.31.73
Brlapi==0.8.3
bs4==0.0.1
cachetools==5.3.2
catalogue==2.0.10
catboost==1.2.2
certifi==2023.5.7
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
colorama==0.4.4
colorful==0.5.5
comm==0.1.4
command-not-found==0.3
confection==0.1.3
contourpy==1.1.1
crcmod==1.7
cryptography==3.4.8
cupshelpers==1.0
cutadapt==3.5
cycler==0.11.0
cymem==2.0.8
Cython==3.0.2
dask==2023.10.0
dask-glm==0.3.0
dask-ml==2023.3.24
datasets==2.14.6
dbus-python==1.2.18
deap==1.4.1
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.7
distributed==2023.10.0
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnaio==0.7.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
evaluate==0.3.0
exceptiongroup==1.1.2
exchange-calendars==4.5
executing==1.2.0
fastai==2.7.13
fastcore==1.5.29
fastdownload==0.0.7
fasteners==0.14.1
fastprogress==1.0.3
filelock==3.12.2
finlab==0.5.0
FinMind==1.6.1
Flask==2.3.3
fonttools==4.42.1
formulaic==0.6.6
frozendict==2.3.8
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.2
gdown==4.7.1
geographiclib==2.0
geopy==2.4.0
gluonts==0.13.7
google-api-core==2.12.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
gpustat==1.1.1
graphviz==0.20.1
greenlet==3.0.0
grpcio==1.50.0
h11==0.14.0
html5lib==1.1
httplib2==0.20.2
huggingface-hub==0.18.0
hyperopt==0.2.7
idna==3.4
imageio==2.31.6
importlib-metadata==6.8.0
iniconfig==2.0.0
interface-meta==1.3.0
ipykernel==6.25.2
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
jeepney==0.7.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
jsonschema==4.17.3
jupyter_client==8.3.1
jupyter_core==5.3.1
keyring==23.5.0
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lifelines==0.27.8
lightgbm==3.3.5
lightning-utilities==0.9.0
lineTool==1.0.3
llvmlite==0.41.1
locket==1.0.0
lockfile==0.12.2
loguru==0.7.2
louis==3.20.0
lxml==4.9.3
lz4==4.3.2
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mlforecast==0.7.3
model-index==0.1.11
monotonic==1.6
more-itertools==8.10.0
mplfinance==0.12.10b0
msgpack==1.0.7
multidict==6.0.4
multipledispatch==1.0.0
multiprocess==0.70.15
multitasking==0.0.11
murmurhash==1.0.10
mysql-connector-python==8.2.0
nest-asyncio==1.5.8
netifaces==0.11.0
networkx==3.2.1
nftables==0.1
nlpaug==1.1.11
nltk==3.8.1
nptyping==2.4.1
numba==0.58.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-ml-py==12.535.108
oauthlib==3.2.0
olefile==0.46
omegaconf==2.2.3
opencensus==0.11.3
opencensus-context==0.1.3
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.28
ordered-set==4.1.0
oss2==2.17.0
outcome==1.2.0
packaging==23.1
panda==0.3.1
pandas==2.1.4
pandas-market-calendars==4.3.2
paramiko==2.9.3
parso==0.8.3
partd==1.4.1
patsy==0.5.3
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.11.0
plotly==5.18.0
pluggy==1.3.0
preshed==3.0.9
prettytable==3.9.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-spy==0.3.14
py4j==0.10.9.7
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycryptodome==3.19.0
pycups==2.0.1
pydantic==1.10.13
pydantic_core==2.6.3
pyecharts==2.0.4
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.7.0
pyluach==2.2.0
pymacaroons==0.13.0
PyMuPDF==1.23.7
PyMuPDFb==1.23.7
PyMySQL==1.1.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
pytesseract==0.3.10
pytest==7.4.2
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
pytorch-lightning==1.9.5
pytorch-metric-learning==1.7.3
pytz==2023.3
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==6.0.1
pyzmq==25.1.1
ray==2.3.1
regex==2023.10.3
reportlab==3.6.8
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.4.2
rsa==4.9
s3transfer==0.7.0
safetensors==0.4.0
scikit-image==0.19.3
scikit-learn==1.2.2
scikit-MDR==0.4.5
scipy==1.11.3
SecretStorage==3.3.1
selenium==4.11.2
sentencepiece==0.1.99
seqeval==1.2.2
simplejson==3.19.1
six==1.16.0
sklearn==0.0.post9
skrebate==0.62
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sparse==0.14.0
SQLAlchemy==2.0.22
sqlparse==0.4.4
srsly==2.4.8
ssh-import-id==5.11
stack-data==0.6.2
statsforecast==1.4.0
statsmodels==0.14.0
stevedore==5.1.0
stopit==1.1.2
systemd-python==234
ta==0.5.25
tabulate==0.9.0
tblib==3.0.0
tenacity==8.2.3
tensorboard==2.15.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
text-unidecode==1.3
texttable==1.6.4
thinc==8.2.1
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.8
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchmetrics==0.11.4
torchvision==0.14.1
tornado==6.3.3
TPOT==0.12.1
tqdm==4.65.2
traitlets==5.10.0
transformers==4.26.1
trio==0.22.2
trio-websocket==0.10.3
typer==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.8.0
unattended-upgrades==0.1
update-checker==0.18.0
urllib3==1.26.18
usb-creator==0.3.7
virtualenv==16.7.10
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wadllib==1.3.6
wasabi==1.1.2
wcwidth==0.2.6
weasel==0.3.3
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.3.7
window-ops==0.0.14
wrapt==1.16.0
wsproto==1.2.0
xdg==5
xgboost==1.7.6
xkit==0.0.0
XlsxWriter==1.2.9
xopen==1.2.1
xxhash==3.4.1
yarl==1.9.2
yfinance==0.2.28
zict==3.0.0
zipp==1.0.0
